requirements.txt
pytest-playwright
pytest==8.3.2
pytest-dotenv==0.5.2
python-dotenv==1.0.1
pytest-html==4.1.1
pytest-xdist==3.6.1
allure-pytest==2.13.5
jsonschema==4.26.0
